Redirect /aniversario/patrocinadores to /patrocinadores
- Move sponsor page content from /aniversario/patrocinadores into
  /patrocinadores (was previously a re-export)
- /aniversario/patrocinadores now client-side redirects to /patrocinadores
- Update nav hrefs in event.ts, PatrocinadoresCta.tsx to point to /patrocinadores
- Remove /aniversario/patrocinadores from sitemap
