+++ src/app/sitemap.ts
@@ -11,7 +11,6 @@ export default function sitemap(): MetadataRoute.Sitemap {
   return [
     { url: `${BASE_URL}/`, lastModified, priority: 1.0 },
     { url: `${BASE_URL}/aniversario`, lastModified, priority: 1.0 },
-    { url: `${BASE_URL}/aniversario/patrocinadores`, lastModified, priority: 0.6 },
     { url: `${BASE_URL}/patrocinadores`, lastModified, priority: 0.6 },
     { url: `${BASE_URL}/codigo-conducta`, lastModified, priority: 0.6 },
     { url: `${BASE_URL}/terminos`, lastModified, priority: 0.6 },
